[fix] Xcode 폴더 에러. 해결
